refactor(region): split the restricted-region surfaces by account-type namespace
